dom/media/webaudio/test/test_decoderDelay.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that decoder delay is handled</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var tests = [
  "half-a-second-8000.mp3",
  "half-a-second-48000.mp3",
];

async function doit(t) {
  var count = 0;
  var context = new OfflineAudioContext(1, 128, 48000);
  tests.forEach(async testfile => {
    var response = await fetch(testfile);
    var buffer = await response.arrayBuffer();
    var decoded = await context.decodeAudioData(buffer);
    is(decoded.duration, 0.5, "The file is half a second.");
    if (++count == tests.length) {
      SimpleTest.finish();
    }
  });
}

doit();

</script>
</pre>
</body>
</html>
